🚚 :: [#670] 프로필 이미지 목록 API 도메인 이동
User -> Image
