+++ Projects/Domains/ImageDomain/Interface/DataSource/RemoteImageDataSource.swift
@@ -3,4 +3,6 @@ import RxSwift
 
 public protocol RemoteImageDataSource {
     func fetchLyricDecoratingBackground() -> Single<[LyricDecoratingBackgroundEntity]>
+    func fetchProfileList() -> Single<[ProfileListEntity]>
+
 }
