+++ Projects/Domains/ImageDomain/Interface/Entity/ProfileListEntity.swift
@@ -1,11 +1,3 @@
-//
-//  ProfileListEntity.swift
-//  DomainModule
-//
-//  Created by KTH on 2023/02/18.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
 import Foundation
 
 public struct ProfileListEntity: Equatable {
