+++ Projects/Domains/ImageDomain/Interface/Repository/ImageRepository.swift
@@ -3,4 +3,6 @@ import RxSwift
 
 public protocol ImageRepository {
     func fetchLyricDecoratingBackground() -> Single<[LyricDecoratingBackgroundEntity]>
+    func fetchProfileList() -> Single<[ProfileListEntity]>
+
 }
