+++ Projects/Domains/ImageDomain/Interface/UseCase/FetchProfileListUseCase.swift
@@ -1,0 +1,6 @@
+import Foundation
+import RxSwift
+
+public protocol FetchProfileListUseCase {
+    func execute() -> Single<[ProfileListEntity]>
+}
