+++ Projects/Domains/ImageDomain/Sources/ResponseDTO/FetchProfileListResponseDTO.swift
@@ -1,13 +1,5 @@
-//
-//  FetchProfileListResponseDTO.swift
-//  DataMappingModule
-//
-//  Created by KTH on 2023/02/18.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
 import Foundation
-import UserDomainInterface
+import ImageDomainInterface
 
 public struct FetchProfileListResponseDTO: Decodable, Equatable {
     public let type: String?
