+++ Projects/Domains/UserDomain/Interface/Repository/UserRepository.swift
@@ -1,17 +1,8 @@
-//
-//  ArtistRepository.swift
-//  DomainModule
-//
-//  Created by KTH on 2023/02/08.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
 import BaseDomainInterface
 import Foundation
 import RxSwift
 
 public protocol UserRepository {
-    func fetchProfileList() -> Single<[ProfileListEntity]>
     func setProfile(image: String) -> Single<BaseEntity>
     func setUserName(name: String) -> Single<BaseEntity>
     func fetchPlayList() -> Single<[PlayListEntity]>
