+++ Projects/Features/MyInfoFeature/Sources/Views/ProfileCollectionViewCell.swift
@@ -1,14 +1,6 @@
-//
-//  ProfileCollectionViewCell.swift
-//  DesignSystem
-//
-//  Created by yongbeomkwak on 2023/01/26.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
 import DesignSystem
 import UIKit
-import UserDomainInterface
+import ImageDomainInterface
 import Utility
 
 public enum FanType: String {
